Create packages/logger; lift shared ConsoleLogger (closes #291)
Creates @transformotion/logger with Logger interface and ConsoleLogger
(structured console output). Both apps now import the shared types and
ConsoleLogger from the package; app-specific AppName union types removed.
createLogger() defaults level to 'info', matching the env-var default.
