+++ MONOREPO.md
@@ -42,6 +42,7 @@ transformotion-apps/
 │   ├── auth-client/                       # Cognito + mock auth service implementations (@transformotion/auth-client)
 │   ├── budget-domain/                     # Budget Tracker domain types and helpers
 │   ├── cache/                             # Shared cache interfaces + MemoryCacheService (@transformotion/cache)
+│   ├── logger/                            # Shared Logger interface + ConsoleLogger (@transformotion/logger)
 │   ├── lambda-middleware/                 # Shared withAuth/withAuthOnly wrappers
 │   ├── runtime-config/                    # Runtime profile + provider resolution (@transformotion/runtime-config)
 │   └── ui/                                # Organisational directory (not itself a package; §3.4)
