+++ apps/stock-analyser/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@transformotion/api-client": "workspace:*",
     "@transformotion/cache": "workspace:*",
+    "@transformotion/logger": "workspace:*",
     "@transformotion/ui-error-boundaries": "workspace:*",
     "@transformotion/ui-primitives": "workspace:*",
     "@transformotion/auth-client": "workspace:*",
