+++ docs/architecture/inventory.md
@@ -47,8 +47,8 @@ The repository contains the following top-level directories:
 
 - `apps/` — four subdirectories: `budget-tracker/`, `launchpad/`,
   `stock-analyser/`, `web/` (0-LOC shell, M3 cleanup)
-- `packages/` — seven subdirectories: `api-client/`, `auth-client/`,
-  `budget-domain/`, `cache/`, `lambda-middleware/`, `runtime-config/`, `ui/`
+- `packages/` — eight subdirectories: `api-client/`, `auth-client/`,
+  `budget-domain/`, `cache/`, `lambda-middleware/`, `logger/`, `runtime-config/`, `ui/`
   (organisational directory; two sub-packages: `ui-error-boundaries`,
   `ui-primitives`). `cycle-engine/` was removed — see Section 1.5.
 - `infrastructure/` — CDK app with `bin/app.ts` entrypoint and
@@ -150,6 +150,13 @@ package. Split landed in M7 / PR #298:
 - `packages/ui/error-boundaries/` (`@transformotion/ui-error-boundaries`) —
   houses `TabErrorBoundary`; 6 vitest tests; 2 consumers
   (`budget-tracker-app.tsx`, `apps/stock-analyser/app/page.tsx`).
+
+`packages/logger/` (`@transformotion/logger`) was created by M7 / PR #291:
+- Exports: `Logger`, `LogLevel`, `LogContext`, `LogEntry`, `LoggerConfig` interfaces; `ConsoleLogger`, `createLogger`, `getLogger`
+- `ConsoleLogger` lifted from `apps/*/lib/services/logger/console-logger.ts` (byte-identical in both apps)
+- `AppName` union type removed from package — `app` fields broadened to `string`; per-app `AppName` types removed
+- `createLogger(app?, level?)` defaults `level` to `'info'` (matches env-var default in both apps)
+- Both apps' `lib/services/logger/index.ts` re-exports from the package; `lib/logging.ts` imports from `@transformotion/logger` directly
 
 `packages/cache/` (`@transformotion/cache`) was created by M7 / PR #290:
 - Exports: `CacheService`, `CacheOptions`, `CacheEntry`, `CacheConfig` interfaces; `MemoryCacheService`, `createCacheService`
